requirements.txt.txt
uvicorn
numpy
tensorflow
fastapi
python-multipart
Pillow